CourseManagementJDBC/src/test/java/packt/book/jee/eclipse/ch5/bean/CourseTest.java
package packt.book.jee.eclipse.ch5.bean;

import org.junit.Assert;
import org.junit.Test;
import packt.book.jee.eclipse.ch4.bean.Course;

public class CourseTest {  
	@Test  public void testIsValidCourse() 
	{    
		Course course = new Course();    
		//First validate without any values set    
		Assert.assertFalse(course.isValidCourse()); 
		//set  name    
		course.setName("course1");    
		Assert.assertFalse(course.isValidCourse());    
		//set zero credits    
		course.setCredits(0);    
		Assert.assertFalse(course.isValidCourse());    
		//now set valid credits    
		course.setCredits(4);    
		Assert.assertTrue(course.isValidCourse());
		//set empty course name
		course.setName("");
		Assert.assertFalse(course.isValidCourse());
	}
}